Regex version of validation assignment
The assignment was to get it working essentially the way I did in the
previous assignment (PHP_robust_validation_assignment) but I went
through and essentially worked on them both at the same time and they
wound up being nearly identical. The part I wanted to show off was the
Index.php (not IndexFINISHED.php) for the extra credit regular
expressions version of the assignment.